+++ index.tsx
@@ -1,2 +1,3 @@
 export { FolderStructure } from "./src/FolderStructure";
 export { SkipLink } from "./src/SkipLink";
+export { Breadcrumbs } from "./src/Breadcrumbs";