+++ src/Breadcrumbs/Breadcrumbs.types.ts
@@ -1,0 +1,8 @@
+export type BreadCrumbProps = React.ComponentPropsWithRef<"nav"> & {
+  breadcrumbs: Array<BreadCrumbType>;
+};
+
+export type BreadCrumbType = {
+  text: string;
+  href: string;
+};