+++ src/SkipLink/SkipLink.types.ts
@@ -1,6 +1,4 @@
-import { ComponentPropsWithRef } from "react";
-
-export interface SkipLinkProps extends ComponentPropsWithRef<"a"> {
+export interface SkipLinkProps extends React.ComponentPropsWithRef<"a"> {
   /**
    * The text to display for the skip link
    * @default 'Skip to main content'